refactor:코드리뷰 반영;파일 위치 변경 및 navigate 추가
