+++ src/common/http.ts
@@ -1,8 +1,6 @@
 import axios from 'axios';
-import { Cookies } from 'react-cookie';
 axios.defaults.xsrfCookieName = 'csrftoken';
 axios.defaults.xsrfHeaderName = 'X-CSRFTOKEN';
-const cookie = new Cookies();
 export const http = axios.create({
   baseURL: 'http://localhost:8000/',
   headers: {
