+++ src/services/Auth/AuthServices.ts
@@ -5,14 +5,12 @@ axios.defaults.withCredentials = true;
 
 const postSignin = (user: User) =>
   http.post(`login`, user, { withCredentials: true });
-const getSignin = () => http.get(`login`);
 const postLogout = () => http.post(`logout`, {}, { withCredentials: true });
 const postSignUp = (user: SignupUser) =>
   http.post(`register`, user, { withCredentials: true });
 
 const AuthServices = {
   postSignin,
-  getSignin,
   postLogout,
   postSignUp,
 };
