+++ src/utils/isLogin.ts
@@ -1,0 +1,1 @@
+export const isLogin = () => !!localStorage.getItem('userId');
